+++ indices/make_range_index/range_index.cc
@@ -3,29 +3,6 @@
 #include <map>
 #include <string>
 
-void Node::Serialize(std::ostream &os) {
-    os.write((char *) &start_time, sizeof(time_t));
-    os.write((char *) &end_time, sizeof(time_t));
-}
-
-Status Node::Deserialize(std::istream &is) {
-    is.read(reinterpret_cast<char *>(&start_time), sizeof(time_t));
-    is.read((char *) &end_time, sizeof(time_t));
-    return Status();
-}
-
-void NodeSummary::Serialize(std::ostream &os){
-    os.write((char *) &start_time, sizeof(time_t));
-    os.write((char *) &end_time, sizeof(time_t));
-}
-
-Status NodeSummary::Deserialize(std::istream &is) {
-    NodeSummary sum;
-    is.read((char *) &sum.start_time, sizeof(time_t));
-    is.read((char *) &sum.end_time, sizeof(time_t));
-    return Status();
-}
-
 Status get_last_updated_and_granularity(
     gcs::Client* client, time_t &granularity, time_t &last_updated,
     std::string index_bucket
